requirements.txt
black
flake8
h5py==3.6.0
isort
nbqa
optuna
pytorch-adapt[ignite,record-keeper]==0.0.73
PyYAML
seaborn
submitit
tensorboard
timm
tqdm

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cu111
torchvision==0.11.2+cu111